+++ src/process/sysdep_DRAGONFLY.c
@@ -158,7 +158,7 @@ int initprocesstree_sysdep(ProcessTree_T **reference, ProcessEngine_Flags pflags
                 pt[i].threads      = pinfo[i].kp_nthreads;
                 pt[i].uptime       = systeminfo.time / 10. - pinfo[i].kp_start.tv_sec;
                 pt[i].cpu.time     = (double)((pinfo[i].kp_lwp.kl_uticks + pinfo[i].kp_lwp.kl_sticks + pinfo[i].kp_lwp.kl_iticks) / 1000000.);
-                pt[i].memory.usage = (unsigned long)(pinfo[i].kp_vm_rssize * pagesize);
+                pt[i].memory.usage = (uint64_t)pinfo[i].kp_vm_rssize * (uint64_t)pagesize;
                 pt[i].zombie       = pinfo[i].kp_stat == SZOMB ? true : false;
                 if (pflags & ProcessEngine_CollectCommandLine) {
                         char **args = kvm_getargv(kvm_handle, &pinfo[i], 0);
@@ -222,7 +222,7 @@ boolean_t used_system_memory_sysdep(SystemInfo_T *si) {
                 LogError("system statistic error -- wired memory usage statics error\n");
                 return false;
         }
-        si->total_mem = (active + wired) * pagesize;
+        si->total_mem = (uint64_t)(active + wired) * (uint64_t)pagesize;
 
         /* Swap */
         unsigned int used;
@@ -231,20 +231,20 @@ boolean_t used_system_memory_sysdep(SystemInfo_T *si) {
                 si->swap_max = 0;
                 return false;
         }
-        si->total_swap = used * pagesize;
+        si->total_swap = (uint64_t)used * (uint64_t)pagesize;
         if (sysctlbyname("vm.swap_cache_use", &used, &len, NULL, 0) == -1) {
                 LogError("system statistic error -- cannot get swap usage: %s\n", STRERROR);
                 si->swap_max = 0;
                 return false;
         }
-        si->total_swap += used * pagesize;
+        si->total_swap += (uint64_t)used * (uint64_t)pagesize;
         unsigned int free;
         if (sysctlbyname("vm.swap_size", &free, &len, NULL, 0) == -1) {
                 LogError("system statistic error -- cannot get swap usage: %s\n", STRERROR);
                 si->swap_max = 0;
                 return false;
         }
-        si->swap_max = free * pagesize + si->total_swap;
+        si->swap_max = (uint64_t)free * (uint64_t)pagesize + si->total_swap;
         return true;
 }
